Address markdown rendering review feedback
