+++ Sources/Panels/MarkdownPointerObserver.swift
@@ -1,0 +1,112 @@
+import AppKit
+import SwiftUI
+
+struct MarkdownPointerObserver: NSViewRepresentable {
+    let onPointerDown: () -> Void
+
+    func makeNSView(context: Context) -> MarkdownPanelPointerObserverView {
+        let view = MarkdownPanelPointerObserverView()
+        view.onPointerDown = onPointerDown
+        return view
+    }
+
+    func updateNSView(_ nsView: MarkdownPanelPointerObserverView, context: Context) {
+        nsView.onPointerDown = onPointerDown
+    }
+}
+
+final class MarkdownPanelPointerObserverView: NSView {
+    var onPointerDown: (() -> Void)?
+    private var eventMonitor: Any?
+    private weak var forwardedMouseTarget: NSView?
+
+    override var mouseDownCanMoveWindow: Bool { false }
+
+    override init(frame frameRect: NSRect) {
+        super.init(frame: frameRect)
+        installEventMonitorIfNeeded()
+    }
+
+    required init?(coder: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
+    }
+
+    deinit {
+        if let eventMonitor {
+            NSEvent.removeMonitor(eventMonitor)
+        }
+    }
+
+    override func hitTest(_ point: NSPoint) -> NSView? {
+        guard PaneFirstClickFocusSettings.isEnabled(),
+              window?.isKeyWindow != true,
+              bounds.contains(point) else { return nil }
+        return self
+    }
+
+    override func acceptsFirstMouse(for event: NSEvent?) -> Bool {
+        PaneFirstClickFocusSettings.isEnabled()
+    }
+
+    override func mouseDown(with event: NSEvent) {
+        onPointerDown?()
+        forwardedMouseTarget = forwardedTarget(for: event)
+        forwardedMouseTarget?.mouseDown(with: event)
+    }
+
+    override func mouseDragged(with event: NSEvent) {
+        forwardedMouseTarget?.mouseDragged(with: event)
+    }
+
+    override func mouseUp(with event: NSEvent) {
+        forwardedMouseTarget?.mouseUp(with: event)
+        forwardedMouseTarget = nil
+    }
+
+    func shouldHandle(_ event: NSEvent) -> Bool {
+        guard event.type == .leftMouseDown,
+              let window,
+              event.window === window,
+              !isHiddenOrHasHiddenAncestor else { return false }
+        if PaneFirstClickFocusSettings.isEnabled(), window.isKeyWindow != true {
+            return false
+        }
+        let point = convert(event.locationInWindow, from: nil)
+        return bounds.contains(point)
+    }
+
+    func handleEventIfNeeded(_ event: NSEvent) -> NSEvent {
+        guard shouldHandle(event) else { return event }
+        DispatchQueue.main.async { [weak self] in
+            self?.onPointerDown?()
+        }
+        return event
+    }
+
+    private func installEventMonitorIfNeeded() {
+        guard eventMonitor == nil else { return }
+        eventMonitor = NSEvent.addLocalMonitorForEvents(matching: [.leftMouseDown]) { [weak self] event in
+            self?.handleEventIfNeeded(event) ?? event
+        }
+    }
+
+    private func forwardedTarget(for event: NSEvent) -> NSView? {
+        guard let window else {
+#if DEBUG
+            NSLog("MarkdownPanelPointerObserverView.forwardedTarget skipped, window=0 contentView=0")
+#endif
+            return nil
+        }
+        guard let contentView = window.contentView else {
+#if DEBUG
+            NSLog("MarkdownPanelPointerObserverView.forwardedTarget skipped, window=1 contentView=0")
+#endif
+            return nil
+        }
+        isHidden = true
+        defer { isHidden = false }
+        let point = contentView.convert(event.locationInWindow, from: nil)
+        let target = contentView.hitTest(point)
+        return target === self ? nil : target
+    }
+}
